Fix uninitialized variable.
	* components/help/converters/gnome-info2html2/main.c: (main):
	Fix uninitialized variable.

	* components/services/vault/command-line/main.c:
	Include stlib.h needed for exit().

	* libnautilus-extensions/nautilus-directory-async.c:
	(call_ready_callbacks):
	Fix uninitialized variable.

	* libnautilus-extensions/nautilus-gdk-pixbuf-extensions.c:
	(nautilus_gdk_pixbuf_draw_text):
	Workaround int vs uint casting issues with new gcc compiler.

	* libnautilus-extensions/nautilus-mime-actions.c:
	(extract_prefix_add_suffix):
	Workaround int vs uint casting issues with new gcc compiler.

+++ libnautilus-private/nautilus-gdk-pixbuf-extensions.c
@@ -549,7 +549,7 @@ nautilus_gdk_pixbuf_draw_text (GdkPixbuf	*pixbuf,
 			/* Extract the next line of text. */
 			end_of_line = strchr (line, '\n');
 			line_length = end_of_line == NULL
-				? strlen (line)
+				? (int) strlen (line)
 				: end_of_line - line;
 			
 			/* Draw the next line of text. */
